requirements.txt
asttokens==2.4.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
drake==1.30.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gradescope-utils==0.5.0
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpld3==0.5.10
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.6
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scipy==1.14.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.0
timeout-decorator==0.5.0
torch==2.3.1
torchvision==0.18.1
traitlets==5.14.3
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
underactuated==2024.6.22
wcwidth==0.2.13
widgetsnbextension==4.0.11
